arm64: fix SetAlarm tick computation truncating the timer multiplier
The absolute tick target was computed with integer division by
uint64(TimerMultiplier), truncating non-integer multipliers (52 vs
52.083 ns/tick at 19.2 MHz): alarms fire late by the truncation ratio
of total counter uptime — hundreds of milliseconds within minutes,
since the counter runs from power-on. Compute in float64 instead.

Hardware-verified on the Arduino UNO Q: WFI wake-up lateness dropped
from ~169 ms average to 2.1 µs.

+++ arm64/timer.go
@@ -94,7 +94,11 @@ func (cpu *CPU) SetAlarm(ns int64) {
 		return
 	}
 
-	set := uint64(ns) / uint64(cpu.TimerMultiplier)
+	// float division: truncating a non-integer multiplier makes the
+	// absolute tick target drift late by the truncation ratio of the
+	// total counter uptime (e.g. 52 vs 52.083 ns/tick at 19.2 MHz is
+	// ~0.16%: hundreds of ms within minutes of counter operation)
+	set := uint64(float64(ns) / cpu.TimerMultiplier)
 	now := read_cntpct()
 	cnt := set - now
